+++ CHANGELOG.md
@@ -5,14 +5,16 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased 3.4](https://github.com/opensearch-project/anomaly-detection/compare/2.x...HEAD)
 ### Features
+- Add agent summary ([#801](https://github.com/opensearch-project/dashboards-flow-framework/pull/801))
+- [Agentic Search] Add MCP server support ([#802](https://github.com/opensearch-project/dashboards-flow-framework/pull/802))
+
 ### Enhancements
 - Clean up / hide complex fields on agent configuration ([#796](https://github.com/opensearch-project/dashboards-flow-framework/pull/796))
-- Add agent summary ([#801](https://github.com/opensearch-project/dashboards-flow-framework/pull/801))
 - Clean up agent summary formatting ([#803](https://github.com/opensearch-project/dashboards-flow-framework/pull/803))
-- [Agentic Search] Add MCP server support ([#802](https://github.com/opensearch-project/dashboards-flow-framework/pull/802))
 - [Agentic Search] Improve export / next steps UX ([#805](https://github.com/opensearch-project/dashboards-flow-framework/pull/805))
 - [Agentic Search] Simplify form inputs ([#807](https://github.com/opensearch-project/dashboards-flow-framework/pull/807))
 - [Agentic Search] Simplify form inputs II ([#808](https://github.com/opensearch-project/dashboards-flow-framework/pull/808))
+- Integrate with memory ([#809](https://github.com/opensearch-project/dashboards-flow-framework/pull/809))
 ### Bug Fixes
 ### Infrastructure
 ### Documentation